+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Library:
 2. [feat] AdbImpl 支持多显示器，可以指定需要操控的显示器 ID。
 3. [feat] 支持 scrcpy 截图与控制方法
 4. [fix] 修复 WindowsImpl 缺少 windll 导入的问题
+5. [feat] **BREAKING** `Device` 类的 `start` 与 `stop` 方法现在会进行线程检查，不可以跨线程调用这两个方法
 
 Devtool：
 1. [feat] 新增最近打开文件功能